# Lugazi-system/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: DCL Lugazi ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only) — reverts them to member-level
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Change a user role in real time (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleChange"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ── Role Upgrade Requests ──
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update a report
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ── Pipeline ──
  /pipeline:
    get:
      operationId: listPipeline
      tags: [pipeline]
      summary: List pipeline contacts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pipeline list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineContact"

    post:
      operationId: createPipelineContact
      tags: [pipeline]
      summary: Add a pipeline contact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineContact"
  /pipeline/{id}:
    patch:
      operationId: updatePipelineContact
      tags: [pipeline]
      summary: Update pipeline contact stage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineContact"

    delete:
      operationId: deletePipelineContact
      tags: [pipeline]
      summary: Delete pipeline contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Documents ──
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, displayName, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        photoUrl:
          type: ["string", "null"]
        branchId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        photoUrl:
          type: string
        phone:
          type: string
        branchId:
          type: integer
        isActive:
          type: boolean
    RoleChange:
      type: object
      required: [role]
      properties:
        role:
          type: string

    # Role Requests
...
    Report:
      type: object
      required: [id, title, type, submittedBy, period, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        submittedBy:
          type: integer
        submittedByName:
          type: ["string", "null"]
        branchId:
          type: ["integer", "null"]
        content:
          type: string
        attendance:
          type: ["integer", "null"]
        soulWinning:
          type: ["integer", "null"]
        period:
          type: string
        status:
          type: string
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [title, type, content, period]
      properties:
        title:
          type: string
        type:
          type: string
        content:
          type: string
        period:
          type: string
        branchId:
          type: integer
        attendance:
          type: integer
        soulWinning:
          type: integer
    ReportUpdate:
      type: object
      properties:
        content:
          type: string
        attendance:
          type: integer
        soulWinning:
          type: integer
        status:
          type: string

    # Pipeline
    PipelineContact:
      type: object
      required: [id, name, phone, stage, assignedTo, source, branchId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        stage:
          type: string
        assignedTo:
          type: integer
        assignedToName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: string
        branchId:
          type: integer
        createdAt:
          type: string
        lastContactedAt:
          type: ["string", "null"]
    PipelineInput:
      type: object
      required: [name, phone, stage, source, branchId]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        stage:
          type: string
        notes:
          type: string
        source:
          type: string
        branchId:
          type: integer
    PipelineUpdate:
      type: object
      properties:
        stage:
          type: string
        notes:
          type: string
        lastContactedAt:
          type: string

    # Documents
...