# Lugazi-system/lib/api-zod/src/generated/types/pipelineContact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DCL Lugazi ERP API
 * OpenAPI spec version: 1.0.0
 */

export interface PipelineContact {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  stage: string;
  assignedTo: number;
  /** @nullable */
  assignedToName?: string | null;
  /** @nullable */
  notes?: string | null;
  source: string;
  branchId: number;
  createdAt: string;
  /** @nullable */
  lastContactedAt?: string | null;
}