# Lugazi-system/lib/api-zod/src/generated/types/pipelineInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DCL Lugazi ERP API
 * OpenAPI spec version: 1.0.0
 */

export interface PipelineInput {
  name: string;
  phone: string;
  email?: string;
  stage: string;
  notes?: string;
  source: string;
  branchId: number;
}